[#IOCIT-34] Small refactory, move client generator from service to client
