+++ src/clients/pn-clients.ts
@@ -1,4 +1,5 @@
 import nodeFetch from "node-fetch";
+import { ValidUrl } from "@pagopa/ts-commons/lib/url";
 import {
   Client,
   createClient
@@ -18,9 +19,13 @@ export function PnAPIClient(
     fetchApi
   });
 }
-
 export type PnAPIClient = typeof PnAPIClient;
 
+/**
+ * Piattaforma Notifiche client for service Activation management.
+ * Accept baseUrl and API key to configure the client generated from
+ * OAS3 specification.
+ */
 export const PnAddressBookIOClient = (
   baseUrl: string,
   apiKey: string,
@@ -37,4 +42,34 @@ export const PnAddressBookIOClient = (
         ApiKeyAuth: apiKey
       })
   });
-export type PnAddressBookClient = typeof PnAddressBookIOClient;
+export type PnAddressBookIOClient = typeof PnAddressBookIOClient;
+
+/**
+ * Available PN environments
+ */
+export enum PNEnvironment {
+  PRODUCTION = "PRODUCTION",
+  UAT = "UAT",
+  DEV = "DEV"
+}
+
+/**
+ * Generate a Client Selector based on enviroment.
+ */
+export const PNClientFactory = (
+  pnApiUrlProd: ValidUrl,
+  pnApiKeyProd: string,
+  pnApiUrlUAT: ValidUrl,
+  pnApiKeyUAT: string,
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+  fetchApi: typeof fetch = (nodeFetch as any) as typeof fetch
+) => (pnEnvironment: PNEnvironment) => {
+  switch (pnEnvironment) {
+    case PNEnvironment.PRODUCTION:
+      return PnAddressBookIOClient(pnApiUrlProd.href, pnApiKeyProd, fetchApi);
+    case PNEnvironment.UAT:
+      return PnAddressBookIOClient(pnApiUrlUAT.href, pnApiKeyUAT, fetchApi);
+    default:
+      throw new Error("Unimplemented PN Environment");
+  }
+};
